Upgrade column types for status/mode.
These could theoretically go higher than tinyint in the future if modes are compounded.

+++ database/migrations/2019_10_23_182119_create_files_table.php
@@ -26,8 +26,8 @@ class CreateFilesTable extends Migration
             $table->json('input_settings')->nullable();
             $table->ipAddress('ip_address');
             $table->string('session_id');
-            $table->unsignedTinyInteger('status');
-            $table->unsignedTinyInteger('mode');
+            $table->unsignedInteger('status');
+            $table->unsignedInteger('mode');
             $table->string('type')->nullable();
             $table->json('columns')->nullable();
             $table->unsignedInteger('column_count');
